dry up command setup and windows assumes powershell for now
